Game Menu
Created the Game class which gives three options: Play chess, view stats, and exit the game. The view stats reads from two .json files (used to be .txt but realized I have them in the .gitignore and plus .json is kind of industry standard. I still need to properly format them as .json files though right now it is unnecessary as the data is minimal and can be read as a text file)